+++ core/src/time_stepping/runge_kutta_third_order.c
@@ -5,7 +5,7 @@
 #include <stdlib.h>
 #include <stdio.h>
 
-int runge_kutta_third_order(State *state_0, State *state_p1, double delta_t, Grid *grid, Dualgrid *dualgrid, short dissipation_on, short rad_on, short add_comps_on)
+int runge_kutta_third_order(State *state_0, State *state_p1, double delta_t, Grid *grid, Dualgrid *dualgrid, int dissipation_on, int rad_on, int add_comps_on)
 {
     State *state_star = malloc(sizeof(State));
     State *tendency_0 = malloc(sizeof(State));
