+++ core/src/time_stepping/time_stepping.h
@@ -1,2 +1,2 @@
-int runge_kutta_third_order(State *, State *, double, Grid *, Dualgrid *, short, short, short);
-int runge_kutta_fourth_order(State *, State *, double, Grid *, Dualgrid *, short, short, short);
+int runge_kutta_third_order(State *, State *, double, Grid *, Dualgrid *, int, int, int);
+int runge_kutta_fourth_order(State *, State *, double, Grid *, Dualgrid *, int, int, int);
